Organising the Repository functions into their own module. Repository is now just a pass-through struct to get to a RepositoryRef, which holds only a reference to a git2::Repository. The git Browser now works with a `RepositoryRef` as opposed to the reference of a `Repository`.
