+++ src/vcs/git.rs
@@ -65,484 +65,30 @@
 // Re-export git2 as sub-module
 pub use git2::{self, BranchType, Error as Git2Error, Oid, Time};
 
+mod repo;
+pub use repo::{History, Repository, RepositoryRef};
+
 pub mod error;
 mod object;
 
-pub use crate::vcs::git::object::*;
+pub use crate::{
+    diff::Diff,
+    vcs::git::object::{Branch, BranchName, Commit, RevObject, Signature, TagName},
+};
+
 use crate::{
-    diff,
-    diff::*,
     file_system,
     file_system::directory,
-    tree::*,
     vcs,
     vcs::{git::error::*, VCS},
 };
 use nonempty::NonEmpty;
-use std::{cmp::Ordering, collections::HashMap, convert::TryFrom, str};
-
-/// A `History` that uses `git2::Commit` as the underlying artifact.
-pub type History = vcs::History<Commit>;
-
-/// Wrapper around the `git2`'s `git2::Repository` type.
-/// This is to to limit the functionality that we can do
-/// on the underlying object.
-pub struct Repository(pub(crate) git2::Repository);
-
-/// OrderedCommit is to allow for us to identify an ordering of commit history
-/// as we enumerate over a revwalk of commits, by assigning each commit an
-/// identifier.
-#[derive(Clone)]
-struct OrderedCommit {
-    id: usize,
-    commit: Commit,
-}
-
-impl std::fmt::Debug for OrderedCommit {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(
-            f,
-            "OrderedCommit {{ id: {}, commit: {} }}",
-            self.id, self.commit.id
-        )
-    }
-}
-
-impl OrderedCommit {
-    fn compare_by_id(&self, other: &Self) -> Ordering {
-        self.id.cmp(&other.id).reverse()
-    }
-}
-
-impl From<OrderedCommit> for Commit {
-    fn from(ordered_commit: OrderedCommit) -> Self {
-        ordered_commit.commit
-    }
-}
-
-impl<'repo> Repository {
-    /// Open a git repository given its URI.
-    ///
-    /// # Errors
-    ///
-    /// * [`error::Error::Git`]
-    pub fn new(repo_uri: &str) -> Result<Self, Error> {
-        git2::Repository::open(repo_uri)
-            .map(Repository)
-            .map_err(Error::from)
-    }
-
-    /// List the branches within a repository, filtering out ones that do not
-    /// parse correctly.
-    ///
-    /// # Errors
-    ///
-    /// * [`error::Error::Git`]
-    pub fn list_branches(&self, filter: Option<BranchType>) -> Result<Vec<Branch>, Error> {
-        self.0
-            .branches(filter)
-            .map_err(Error::from)
-            .and_then(|mut branches| {
-                branches.try_fold(vec![], |mut acc, branch| {
-                    let (branch, branch_type) = branch?;
-                    let name = BranchName::try_from(branch.name_bytes()?)?;
-                    let branch = Branch {
-                        name,
-                        locality: branch_type,
-                    };
-                    acc.push(branch);
-                    Ok(acc)
-                })
-            })
-    }
-
-    /// List the tags within a repository, filtering out ones that do not parse
-    /// correctly.
-    ///
-    /// # Errors
-    ///
-    /// * [`error::Error::Git`]
-    pub fn list_tags(&self) -> Result<Vec<TagName>, Error> {
-        let tags = self.0.tag_names(None)?;
-        Ok(tags
-            .into_iter()
-            .filter_map(|tag| tag.map(TagName::new))
-            .collect())
-    }
-
-    /// Create a [`RevObject`] given a
-    /// [`revspec`](https://git-scm.com/docs/git-rev-parse#_specifying_revisions) string.
-    ///
-    /// # Errors
-    ///
-    /// * [`error::Error::Git`]
-    /// * [`error::Error::RevParseFailure`]
-    pub fn rev(&self, spec: &str) -> Result<RevObject, Error> {
-        RevObject::from_revparse(&self.0, spec)
-    }
-
-    /// Create a [`History`] given a
-    /// [`revspec`](https://git-scm.com/docs/git-rev-parse#_specifying_revisions) string.
-    ///
-    /// # Errors
-    ///
-    /// * [`error::Error::Git`]
-    /// * [`error::Error::RevParseFailure`]
-    pub fn revspec(&self, spec: &str) -> Result<History, Error> {
-        let rev = self.rev(spec)?;
-        let commit = rev.into_commit(&self.0)?;
-        self.commit_to_history(commit)
-    }
-
-    /// Get the [`Diff`] between two commits.
-    pub fn diff(&self, from: &'repo git2::Commit, to: &'repo git2::Commit) -> Result<Diff, Error> {
-        use git2::{Delta, Patch};
-
-        let mut diff = Diff::new();
-        let git_diff = self.diff_commits(from, Some(to))?;
-
-        for (idx, delta) in git_diff.deltas().enumerate() {
-            match delta.status() {
-                Delta::Added => {
-                    let diff_file = delta.new_file();
-                    let path = diff_file.path().ok_or(diff::git::Error::PathUnavailable)?;
-                    let path = file_system::Path::try_from(path.to_path_buf())?;
-
-                    diff.add_created_file(path);
-                },
-                Delta::Deleted => {
-                    let diff_file = delta.old_file();
-                    let path = diff_file.path().ok_or(diff::git::Error::PathUnavailable)?;
-                    let path = file_system::Path::try_from(path.to_path_buf())?;
-
-                    diff.add_deleted_file(path);
-                },
-                Delta::Modified => {
-                    let diff_file = delta.new_file();
-                    let path = diff_file.path().ok_or(diff::git::Error::PathUnavailable)?;
-                    let path = file_system::Path::try_from(path.to_path_buf())?;
-
-                    let patch = Patch::from_diff(&git_diff, idx)?;
-
-                    if let Some(patch) = patch {
-                        let mut hunks: Vec<Hunk> = Vec::new();
-
-                        for h in 0..patch.num_hunks() {
-                            let (hunk, hunk_lines) = patch.hunk(h)?;
-                            let header = hunk.header().to_owned();
-                            let mut lines: Vec<LineDiff> = Vec::new();
-
-                            for l in 0..hunk_lines {
-                                let line = patch.line_in_hunk(h, l)?;
-                                let line = LineDiff::try_from(line)?;
-                                lines.push(line);
-                            }
-                            hunks.push(Hunk { header, lines });
-                        }
-                        diff.add_modified_file(path, hunks);
-                    } else if diff_file.is_binary() {
-                        diff.add_modified_binary_file(path);
-                    } else {
-                        return Err(diff::git::Error::PatchUnavailable(path).into());
-                    }
-                },
-                Delta::Renamed => {
-                    let old = delta
-                        .old_file()
-                        .path()
-                        .ok_or(diff::git::Error::PathUnavailable)?;
-                    let new = delta
-                        .new_file()
-                        .path()
-                        .ok_or(diff::git::Error::PathUnavailable)?;
-
-                    let old_path = file_system::Path::try_from(old.to_path_buf())?;
-                    let new_path = file_system::Path::try_from(new.to_path_buf())?;
-
-                    diff.add_moved_file(old_path, new_path);
-                },
-                Delta::Copied => {
-                    let old = delta
-                        .old_file()
-                        .path()
-                        .ok_or(diff::git::Error::PathUnavailable)?;
-                    let new = delta
-                        .new_file()
-                        .path()
-                        .ok_or(diff::git::Error::PathUnavailable)?;
-
-                    let old_path = file_system::Path::try_from(old.to_path_buf())?;
-                    let new_path = file_system::Path::try_from(new.to_path_buf())?;
-
-                    diff.add_copied_file(old_path, new_path);
-                },
-                status => {
-                    return Err(diff::git::Error::DeltaUnhandled(status).into());
-                },
-            }
-        }
-
-        Ok(diff)
-    }
-
-    /// Get a particular `Commit`.
-    pub(crate) fn get_commit(&'repo self, oid: Oid) -> Result<git2::Commit<'repo>, Error> {
-        let commit = self.0.find_commit(oid)?;
-        Ok(commit)
-    }
-
-    /// Build a [`History`] using the `head` reference.
-    pub(crate) fn head(&'repo self) -> Result<History, Error> {
-        let head = self.0.head()?;
-        self.to_history(&head)
-    }
-
-    /// Turn a [`git2::Reference`] into a [`History`] by completing
-    /// a revwalk over the first commit in the reference.
-    pub(crate) fn to_history(
-        &'repo self,
-        history: &git2::Reference<'repo>,
-    ) -> Result<History, Error> {
-        let head = history.peel_to_commit()?;
-        self.commit_to_history(head)
-    }
-
-    /// Turn a [`git2::Reference`] into a [`History`] by completing
-    /// a revwalk over the first commit in the reference.
-    pub(crate) fn commit_to_history(&'repo self, head: git2::Commit) -> Result<History, Error> {
-        let head_id = head.id();
-        let mut commits = NonEmpty::new(Commit::try_from(head)?);
-        let mut revwalk = self.0.revwalk()?;
-
-        // Set the revwalk to the head commit
-        revwalk.push(head_id)?;
-
-        for commit_result_id in revwalk {
-            // The revwalk iter returns results so
-            // we unpack these and push them to the history
-            let commit_id: Oid = commit_result_id?;
-
-            // Skip the head commit since we have processed it
-            if commit_id == head_id {
-                continue;
-            }
-
-            let commit = Commit::try_from(self.0.find_commit(commit_id)?)?;
-            commits.push(commit);
-        }
-
-        Ok(vcs::History(commits))
-    }
-
-    /// Extract the signature from a commit
-    ///
-    /// # Arguments
-    ///
-    /// `commit_oid` - The object ID of the commit
-    /// `field` - the name of the header field containing the signature block;
-    ///           pass `None` to extract the default 'gpgsig'
-    pub(crate) fn extract_signature(
-        &'repo self,
-        commit_oid: &Oid,
-        field: Option<&str>,
-    ) -> Result<Option<Signature>, Error> {
-        // Match is necessary here because according to the documentation for
-        // git_commit_extract_signature at
-        // https://libgit2.org/libgit2/#HEAD/group/commit/git_commit_extract_signature
-        // the return value for a commit without a signature will be GIT_ENOTFOUND
-        match self.0.extract_signature(commit_oid, field) {
-            Err(error) => {
-                if error.code() == git2::ErrorCode::NotFound {
-                    Ok(None)
-                } else {
-                    Err(error.into())
-                }
-            },
-            Ok(sig) => Ok(Some(Signature::from_buf(sig.0))),
-        }
-    }
-
-    pub(crate) fn revision_branches(&self, oid: &Oid) -> Result<Vec<Branch>, Error> {
-        let branches = self
-            .0
-            .branches(Some(BranchType::Local))?
-            .collect::<Result<Vec<(git2::Branch, BranchType)>, git2::Error>>()?;
-
-        let mut contained_branches = vec![];
-
-        branches.into_iter().try_for_each(|(branch, locality)| {
-            self.reachable_from(&branch.get(), &oid)
-                .and_then(|contains| {
-                    if contains {
-                        let branch = Branch::from_git_branch(branch, locality)?;
-                        contained_branches.push(branch);
-                    }
-                    Ok(())
-                })
-        })?;
-
-        Ok(contained_branches)
-    }
-
-    fn reachable_from(&self, reference: &git2::Reference, oid: &Oid) -> Result<bool, Error> {
-        let other = reference.peel_to_commit()?.id();
-        let is_descendant = self.0.graph_descendant_of(other, *oid)?;
-
-        Ok(other == *oid || is_descendant)
-    }
-
-    /// Get the history of the file system where the head of the [`NonEmpty`] is
-    /// the latest commit.
-    fn file_history(
-        &'repo self,
-        commit: Commit,
-    ) -> Result<Forest<file_system::Label, NonEmpty<OrderedCommit>>, Error> {
-        let mut file_histories = Forest::root();
-        self.collect_file_history(&commit.id, &mut file_histories)?;
-        Ok(file_histories)
-    }
-
-    fn collect_file_history(
-        &'repo self,
-        commit_id: &Oid,
-        file_histories: &mut Forest<file_system::Label, NonEmpty<OrderedCommit>>,
-    ) -> Result<(), Error> {
-        let mut revwalk = self.0.revwalk()?;
-
-        // Set the revwalk to the head commit
-        revwalk.push(*commit_id)?;
-
-        for (id, commit_result) in revwalk.enumerate() {
-            let parent_id = commit_result?;
-
-            let parent = self.0.find_commit(parent_id)?;
-            let paths = self.diff_commit_and_parents(&parent)?;
-            let parent_commit = Commit::try_from(parent)?;
-            for path in paths {
-                let parent_commit = OrderedCommit {
-                    id,
-                    commit: parent_commit.clone(),
-                };
-
-                file_histories.insert_with(
-                    path.0,
-                    NonEmpty::new(parent_commit.clone()),
-                    |commits| commits.push(parent_commit),
-                );
-            }
-        }
-        Ok(())
-    }
-
-    fn diff_commit_and_parents(
-        &'repo self,
-        commit: &'repo git2::Commit,
-    ) -> Result<Vec<file_system::Path>, Error> {
-        let mut parents = commit.parents();
-        let head = parents.next();
-        let mut touched_files = vec![];
-
-        let mut add_deltas = |diff: git2::Diff| -> Result<(), Error> {
-            let deltas = diff.deltas();
-
-            for delta in deltas {
-                let new = delta.new_file().path().ok_or(Error::LastCommitException)?;
-                let path = file_system::Path::try_from(new.to_path_buf())?;
-                touched_files.push(path);
-            }
-
-            Ok(())
-        };
-
-        match head {
-            None => {
-                let diff = self.diff_commits(&commit, None)?;
-                add_deltas(diff)?;
-            },
-            Some(parent) => {
-                let diff = self.diff_commits(&commit, Some(&parent))?;
-                add_deltas(diff)?;
-
-                for parent in parents {
-                    let diff = self.diff_commits(&commit, Some(&parent))?;
-                    add_deltas(diff)?;
-                }
-            },
-        }
-
-        Ok(touched_files)
-    }
-
-    fn diff_commits(
-        &'repo self,
-        left: &'repo git2::Commit,
-        right: Option<&'repo git2::Commit>,
-    ) -> Result<git2::Diff, Error> {
-        let left_tree = left.tree()?;
-        let right_tree = right.map_or(Ok(None), |commit| commit.tree().map(Some))?;
-
-        let diff = self
-            .0
-            .diff_tree_to_tree(Some(&left_tree), right_tree.as_ref(), None)?;
-
-        Ok(diff)
-    }
-}
-
-impl vcs::GetVCS<Error> for Repository {
-    type RepoId = String;
-
-    fn get_repo(repo_id: Self::RepoId) -> Result<Self, Error> {
-        git2::Repository::open(&repo_id)
-            .map(Repository)
-            .map_err(Error::from)
-    }
-}
-
-impl From<git2::Repository> for Repository {
-    fn from(repo: git2::Repository) -> Self {
-        Repository(repo)
-    }
-}
-
-impl VCS<Commit, Error> for Repository {
-    type HistoryId = String;
-    type ArtefactId = Oid;
-
-    fn get_history(&self, history_id: Self::HistoryId) -> Result<History, Error> {
-        self.revspec(&history_id)
-    }
-
-    fn get_histories(&self) -> Result<Vec<History>, Error> {
-        self.0
-            .references()
-            .map_err(Error::from)
-            .and_then(|mut references| {
-                references.try_fold(vec![], |mut acc, reference| {
-                    reference.map_err(Error::from).and_then(|r| {
-                        let history = self.to_history(&r)?;
-                        acc.push(history);
-                        Ok(acc)
-                    })
-                })
-            })
-    }
-
-    fn get_identifier(artifact: &Commit) -> Self::ArtefactId {
-        artifact.id
-    }
-}
-
-impl std::fmt::Debug for Repository {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, ".git")
-    }
-}
+use std::{collections::HashMap, convert::TryFrom, str};
 
 /// A [`crate::vcs::Browser`] that uses [`Repository`] as the underlying
 /// repository backend, [`git2::Commit`] as the artifact, and [`Error`] for
 /// error reporting.
-pub type Browser<'a> = vcs::Browser<'a, Repository, Commit, Error>;
+pub type Browser<'a> = vcs::Browser<RepositoryRef<'a>, Commit, Error>;
 
 impl<'a> Browser<'a> {
     /// Create a new browser to interact with.
@@ -567,9 +113,10 @@ impl<'a> Browser<'a> {
     /// # }
     /// ```
     pub fn new(repository: &'a Repository) -> Result<Self, Error> {
+        let repository = repository.as_ref();
         let history = repository.head()?;
-        let snapshot = Box::new(|repository: &Repository, history: &History| {
-            let tree = Self::get_tree(&repository.0, history.0.first())?;
+        let snapshot = Box::new(|repository: &RepositoryRef<'a>, history: &History| {
+            let tree = Self::get_tree(&repository.repo_ref, history.0.first())?;
             Ok(directory::Directory::from_hash_map(tree))
         });
         Ok(vcs::Browser {
@@ -597,6 +144,7 @@ impl<'a> Browser<'a> {
     /// # fn main() -> Result<(), Box<dyn Error>> {
     /// let repo = Repository::new("./data/git-platinum")?;
     /// let first_branch = repo
+    ///     .as_ref()
     ///     .list_branches(None)?
     ///     .first()
     ///     .cloned()
@@ -610,9 +158,10 @@ impl<'a> Browser<'a> {
         repository: &'a Repository,
         branch_name: BranchName,
     ) -> Result<Self, Error> {
+        let repository = repository.as_ref();
         let history = repository.get_history(branch_name.name().to_string())?;
-        let snapshot = Box::new(|repository: &Repository, history: &History| {
-            let tree = Self::get_tree(&repository.0, history.0.first())?;
+        let snapshot = Box::new(|repository: &RepositoryRef<'a>, history: &History| {
+            let tree = Self::get_tree(&repository.repo_ref, history.0.first())?;
             Ok(directory::Directory::from_hash_map(tree))
         });
         Ok(vcs::Browser {
@@ -712,7 +261,7 @@ impl<'a> Browser<'a> {
         let name = branch_name.name();
         let is_branch = self
             .repository
-            .0
+            .repo_ref
             .resolve_reference_from_short_name(name)
             .map(|reference| reference.is_branch() || reference.is_remote())?;
 
@@ -769,7 +318,7 @@ impl<'a> Browser<'a> {
 
         if !self
             .repository
-            .0
+            .repo_ref
             .resolve_reference_from_short_name(name)?
             .is_tag()
         {
@@ -878,10 +427,15 @@ impl<'a> Browser<'a> {
     /// [`revspec`](#method.revspec) would be a more common function to use.
     pub fn rev(&mut self, rev: RevObject) -> Result<(), Error> {
         let repository = &self.repository;
-        let commit = rev.into_commit(&repository.0)?;
+        let commit = rev.into_commit(&repository.repo_ref)?;
         let history = repository.commit_to_history(commit)?;
         self.set(history);
         Ok(())
+    }
+
+    /// Get the [`Diff`] between two commits.
+    pub fn diff(&self, from: &'a git2::Commit, to: &'a git2::Commit) -> Result<Diff, Error> {
+        self.repository.diff(from, to)
     }
 
     /// List the names of the _branches_ that are contained in the underlying
@@ -1009,10 +563,12 @@ impl<'a> Browser<'a> {
         let file_history = self.repository.file_history(self.get().first().clone())?;
 
         Ok(file_history.find(path.0).map(|tree| {
-            tree.maximum_by(&|c: &NonEmpty<OrderedCommit>, d| c.first().compare_by_id(&d.first()))
-                .first()
-                .commit
-                .clone()
+            tree.maximum_by(&|c: &NonEmpty<repo::OrderedCommit>, d| {
+                c.first().compare_by_id(&d.first())
+            })
+            .first()
+            .commit
+            .clone()
         }))
     }
 
@@ -1105,7 +661,7 @@ impl<'a> Browser<'a> {
     /// ```
     pub fn revision_branches(&self, revspec: &str) -> Result<Vec<Branch>, Error> {
         let rev = self.repository.rev(revspec)?;
-        let commit = rev.into_commit(&self.repository.0)?;
+        let commit = rev.into_commit(&self.repository.repo_ref)?;
         self.repository.revision_branches(&commit.id())
     }
 
@@ -1438,23 +994,27 @@ mod tests {
         }
     }
 
-    #[test]
-    fn test_diff() {
-        use file_system::*;
-        use pretty_assertions::assert_eq;
+    #[cfg(test)]
+    mod diff {
+        use crate::{diff::*, vcs::git::*};
 
-        let repo = Repository::new("./data/git-platinum").unwrap();
+        #[test]
+        fn test_diff() -> Result<(), Error> {
+            use file_system::*;
+            use pretty_assertions::assert_eq;
 
-        let commit = repo
-            .0
-            .find_commit(Oid::from_str("80bacafba303bf0cdf6142921f430ff265f25095").unwrap())
-            .unwrap();
-        let parent = commit.parent(0).unwrap();
+            let repo = Repository::new("./data/git-platinum")?;
+            let commit = repo
+                .0
+                .find_commit(Oid::from_str("80bacafba303bf0cdf6142921f430ff265f25095")?)
+                .unwrap();
+            let parent = commit.parent(0)?;
 
-        let diff = repo.diff(&parent, &commit).unwrap();
+            let bro = Browser::new(&repo)?;
 
-        assert_eq!(
-            Diff {
+            let diff = bro.diff(&parent, &commit)?;
+
+            let expected_diff = Diff {
                 created: vec![],
                 deleted: vec![],
                 moved: vec![],
@@ -1472,8 +1032,10 @@ mod tests {
                         }]
                     )
                 }]
-            },
-            diff
-        );
+            };
+            assert_eq!(expected_diff, diff);
+
+            Ok(())
+        }
     }
 }
